refactor(http): split http/app.rs into focused submodules
Break the 2071-line src/http/app.rs into a directory module:

- mod.rs      — AppState, GatewayRequestStore, router building, Unix serving,
                and the ACI-headers middleware
- handlers.rs — axum route handlers + query types
- proxy.rs    — middleware UDS forwarding + response finalization
- backend.rs  — direct-to-backend forwarding + upstream response shaping
- util.rs     — header/host helpers, forward-candidate building, guards,
                and error-response builders

Pure relocation: no logic changes. Public router/serve API preserved (still
re-exported by http/mod.rs); cross-module items marked pub(super).
cargo test --all-targets: 220 passed, 0 failed (unchanged from baseline).
